Updated some code to make controls better

+++ src/main/java/me/fishydarwin/fractalexplorer/view/component/JFractalRenderer.java
@@ -29,6 +29,7 @@ public class JFractalRenderer extends JPanel {
         imagePanel = new JImagePanel();
         imagePanel.setBackground(new Color(42, 42, 42));
         add(imagePanel);
+        // TODO: right click save rendered frame
     }
 
     private int maxIterations = 100;
@@ -41,8 +42,8 @@ public class JFractalRenderer extends JPanel {
         this.maxIterations = maxIterations;
     }
 
-    private int offsetX = 0;
-    private int offsetY = 0;
+    private double offsetX = 0;
+    private double offsetY = 0;
 
     private double zoomScale = 1;
 
@@ -120,19 +121,19 @@ public class JFractalRenderer extends JPanel {
 
     }
 
-    public int getOffsetX() {
+    public double getOffsetX() {
         return offsetX;
     }
 
-    public void setOffsetX(int offsetX) {
+    public void setOffsetX(double offsetX) {
         this.offsetX = offsetX;
     }
 
-    public int getOffsetY() {
+    public double getOffsetY() {
         return offsetY;
     }
 
-    public void setOffsetY(int offsetY) {
+    public void setOffsetY(double offsetY) {
         this.offsetY = offsetY;
     }
